backend/app/services/embedding_service.py
import logging
import numpy as np
from typing import List, Dict, Any, Optional
import uuid

# In a real implementation, you would use a proper embedding model
# For this prototype, we'll use a simple mock embedding function
# In production, you would use something like:
# from transformers import AutoTokenizer, AutoModel

from app.services.vector_store_service import vector_store

# Setup logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class EmbeddingService:
    """
    Service for generating and managing embeddings for code chunks.
    """
    
    def __init__(self):
        """
        Initialize the embedding service.
        """
        # In a real implementation, you would load the model here
        # self.tokenizer = AutoTokenizer.from_pretrained("microsoft/codebert-base")
        # self.model = AutoModel.from_pretrained("microsoft/codebert-base")
        logger.info("Embedding service initialized")
    
    def generate_embedding(self, text: str) -> List[float]:
        """
        Generate an embedding for a text.
        
        Args:
            text: The text to embed
            
        Returns:
            A list of floats representing the embedding
        """
        # In a real implementation, you would use the model to generate embeddings
        # Here we'll just create a simple mock embedding
        # The dimensionality is set to 768 to match CodeBERT's output
        
        # Create a deterministic but unique embedding based on the text
        # This is just for demonstration purposes
        text_hash = hash(text) % 10000
        np.random.seed(text_hash)
        
        # Generate a random embedding vector (768 dimensions)
        embedding = np.random.normal(0, 1, 768).tolist()
        
        return embedding
    
    def process_chunks(self, chunks: List[Dict[str, Any]], repo_id: str) -> None:
        """
        Process a list of code chunks, generate embeddings, and store them.
        
        Args:
            chunks: List of code chunks
            repo_id: Repository ID
        """
        if not chunks:
            logger.warning("No chunks to process")
            return
        
        try:
            ids = []
            embeddings = []
            metadatas = []
            documents = []
            
            for chunk in chunks:
                # Generate a unique ID for the chunk
                chunk_id = str(uuid.uuid4())
                
                # Prepare the text for embedding
                text_to_embed = self._prepare_text_for_embedding(chunk)
                
                # Generate the embedding
                embedding = self.generate_embedding(text_to_embed)
                
                # Prepare metadata
                metadata = {
                    'repository_id': repo_id,
                    'file_path': chunk['file_path'],
                    'type': chunk['type'],
                    'name': chunk['name'],
                    'line_range': f"{chunk['line_range'][0]}-{chunk['line_range'][1]}"
                }
                
                if chunk['type'] in ['method', 'function']:
                    metadata['params'] = str(chunk.get('params', []))
                    metadata['returns'] = str(chunk.get('returns', None))
                
                if chunk['type'] == 'method':
                    metadata['class_name'] = chunk['class_name']
                
                # Add to lists
                ids.append(chunk_id)
                embeddings.append(embedding)
                metadatas.append(metadata)
                documents.append(chunk['content'])
            
            # Store in vector database
            vector_store.add_embeddings(ids, embeddings, metadatas, documents)
            logger.info(f"Processed and stored {len(chunks)} chunks for repository {repo_id}")
            
        except Exception as e:
            logger.error(f"Error processing chunks: {str(e)}")
            raise
    
    def _prepare_text_for_embedding(self, chunk: Dict[str, Any]) -> str:
        """
        Prepare a chunk's text for embedding.
        
        Args:
            chunk: The code chunk
            
        Returns:
            Prepared text
        """
        # Combine relevant information for better embeddings
        parts = []
        
        # Add type and name
        parts.append(f"{chunk['type']}: {chunk['name']}")
        
        # Add class name for methods
        if chunk['type'] == 'method':
            parts.append(f"class: {chunk['class_name']}")
        
        # Add docstring if available
        if chunk.get('docstring'):
            parts.append(f"docstring: {chunk['docstring']}")
        
        # Add parameters for methods and functions
        if chunk['type'] in ['method', 'function'] and chunk.get('params'):
            params_str = ", ".join([f"{p['name']}: {p.get('type', 'any')}" for p in chunk['params']])
            parts.append(f"parameters: {params_str}")
        
        # Add return type if available
        if chunk['type'] in ['method', 'function'] and chunk.get('returns'):
            parts.append(f"returns: {chunk['returns']}")
        
        # Add the actual code content
        parts.append(f"code: {chunk['content']}")
        
        return "\n".join(parts)

# Create a singleton instance
embedding_service = EmbeddingService()